// frontend/src/lib/api/api.ts
export interface UpdateUserRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'firstname'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'lastname'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'salt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'photo_url'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateUserRequest
     */
    'disabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'disabled_reason'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    'group'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateUserRequest
     */
    'devices'?: Array<string>;
    /**
     * 
     * @type {Array<number>}
     * @memberof UpdateUserRequest
     */
    'keys'?: Array<number>;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateUserRequest
     */
    'audit'?: Array<string>;
}